The filter code was refactored. Subsequently, improvements were made to the visualization of the filter controls, adjusting layout, colors, and spacing. Another focus was correcting the behavior of multiple filters, ensuring that the system correctly identified active datasets and applied selection logic consistently during the filtering process within the Filters tab. Various other functionalities were also configured to enhance the interface’s fluidity during filtering. Finally, the method for calculating the frequency of filter options was updated, making the filter calculations more meaningful.
Part 2
Enhancements were introduced to the interaction with the application’s charts. The first of these was the setup of two-point selection, which allows users to highlight and work with data intervals or pairs—laying the groundwork for function-fitting in a subsequent task. Next, the axes were configured, with each axis’s description adjusted to ensure a clearer and more proportional visualization. Finally, chart interaction was improved by completing the configuration so that only the selected dataset is displayed when a filter is applied.